Override get_instance with efficient Openstack variant
Problem: get_instance is fetching all instances which is misleading
Solution: in get_instance fetch a single instance

The Openstack driver supports fetching a single instance so override
get_instance with the efficient version.

If your driver is scoped to a users identity then the previous behavior of
fetching all instances and filtering to a specific instance is okay, because
--all instances-- isn't that many. If it's the admin driver and there are
1500+ instances (on the IU provider) that will result in response times on the
order of 10s of seconds, which isn't good. This caused timeout errors in rtwo.

http://libcloud.apache.org/apidocs/0.7.1/libcloud.compute.drivers.openstack.OpenStackNodeDriver.html#ex_get_node_details