Fix buggy refactor of ex_list_all_instances
In 5f510aabefe0d90879, we introduced pagination but also changed a fundamental
behavior. Atmosphere still depends on the fact that ex_list_all_instances
behaves differently for "atmoadmin" than for a regular tenant, so it was not
valid to break that assumption. This was discovered during the j7m v32
release.