Fix pagination error causing instances to be enddated
Previously, we would keep fetching the next page, while the # returned were
500. We inferred that if the # returned were less than the page size, then it
had fetched the last page. (We tested this and it seemed to be reliable).
However, the api sometimes returns 499 even when there are more pages to fetch.
The code has been corrected to fetch until no results are returned, which will
inevitably incur 1 unnecessary api call each time.