Handle duplicates in pagination
In the case where the api returns duplicates, we log an error and fallback to
fetching without pagination.

This handles several problems that arised. In one case a nova service was
really confused about pagination and kept returning the same results in a
different order. Previously, we would loop forever, causing the owner process
to hang indefinitely. This occurred reliably with the xoberns user on the IU
provider.

Another problem (albeit hypothetical) was that we didn't do any checking of
duplicates, so we were possibly returning a non unique list.