Problem: Not handling instance started from deleted volume
When we look for the volume that a server was started from, and the
volume no longer exists, Libcloud throws an exception.

This then results in #39: Instances started from missing volume halts
`monitor_instances_for` task

Solution: Catch exception when getting source volume, returning `None`
for the source, instead of bubbling the exception up further.

Note: Atmosphere will fail if an instance has a source of `None`, but in
this case it shouldn't matter. This scenario can only happen when a
server was started outside of Atmosphere (via API) and so should be
filtered out of instances that Atmosphere cares about before it tries to
create an Atmosphere InstanceSource object.